Move register_token_network into ContractDeployer
talking to a TokenNetworkRegistor contract is quite similar to
deploying a contract. ContractDeployer is the natural place for
register_token_network for now.
