+++ raiden_contracts/tests/deprecation_switch_testnet.py
@@ -11,7 +11,7 @@ from raiden_contracts.constants import (
     MAX_ETH_CHANNEL_PARTICIPANT,
     MAX_ETH_TOKEN_NETWORK,
 )
-from raiden_contracts.deploy.__main__ import register_token_network, setup_ctx
+from raiden_contracts.deploy.__main__ import setup_ctx
 from raiden_contracts.utils.transaction import check_successful_tx
 
 log = getLogger(__name__)
@@ -146,8 +146,7 @@ def deprecation_test_setup(
     assert token_contract.functions.balanceOf(deployer.owner).call() >= token_amount
 
     abi = deployer.contract_manager.get_contract_abi(CONTRACT_TOKEN_NETWORK_REGISTRY)
-    token_network_address = register_token_network(
-        deployer=deployer,
+    token_network_address = deployer.register_token_network(
         token_registry_abi=abi,
         token_registry_address=deployed_contracts[CONTRACT_TOKEN_NETWORK_REGISTRY]['address'],
         token_address=token_address,
